+++ app/(tabs)/loan-amount.tsx
